refactor(vue): Created VuePageLoader and changed directory structure.
